frame_delay_for_stable_fps correction
emcc main.cpp game_engine.cpp vector2d.cpp camera.cpp game.cpp viewport.cpp scene_manager.cpp collider2d.cpp entity.cpp transform_component.cpp tank_entity.cpp sprite_component.cpp render_system.cpp tank_input_component.cpp tank_input_system.cpp rigid_body2d_component.cpp physics_system.cpp rectangle_collider2d.cpp convex_polygon_collider2d.cpp collider2d_collection_component.cpp tileset.cpp tile_entity.cpp fps_entity.cpp fps_system.cpp fps_component.cpp target_viewport_component.cpp render_utils.cpp pugixml.cpp target_viewport.cpp zoom_input_component.cpp zoom_input_system.cpp bullet_entity.cpp unified_tiles_entity.cpp collision_system.cpp point_collider2d.cpp properties_component.cpp rectangle2d.cpp -o ../emscripten_build/tank2.html --use-preload-plugins --preload-file ../assets/dirt_1024_dark_borders.png@dirt_1024_dark_borders.png --preload-file ../assets/aris-first-map.tmx@aris-first-map.tmx --preload-file ../assets/4kmap.tmx@4kmap.tmx --preload-file ../assets/dirt_1024_dark_borders.tsx@dirt_1024_dark_borders.tsx --preload-file ../assets/Mixed_Dirt_1024_dark_borders2x256.tsx@Mixed_Dirt_1024_dark_borders2x256.tsx --preload-file ../assets/Roboto-Regular.ttf@Roboto-Regular.ttf --preload-file ../assets/tank_133x50.png@tank_133x50.png --preload-file ../assets/grey_300x200.png@grey_300x200.png --preload-file ../assets/bullet_w65h20.png@bullet_w65h20.png -s USE_SDL=2 -s USE_SDL_IMAGE=2 -s USE_SDL_TTF=2 -s SDL2_IMAGE_FORMATS='["BMP","GIF","JPEG"]' -s ALLOW_MEMORY_GROWTH=1 --no-heap-copy -s WASM=1 -std=c++14 -O3

+++ src/viewport.cpp
@@ -45,8 +45,8 @@ void ViewPort::FollowEntity(const TransformComponent &transformComponent, const 
 {   
     //calculations do not need rounding because it causes the entity on camera to shake
     static double x,y;
-    x = ( ((transformComponent.Position.x * viewport.entityScale.x) + (static_cast<double>(spriteComponent.sourceRectangle.w * viewport.entityScale.x) / 2.0) ) - (static_cast<double>(allCameras[viewport.cameraID].frame.getW()) / 2.0));
-    y = ( ((transformComponent.Position.y * viewport.entityScale.y) + (static_cast<double>(spriteComponent.sourceRectangle.h * viewport.entityScale.y) / 2.0) ) - (static_cast<double>(allCameras[viewport.cameraID].frame.getH()) / 2.0));
+    x = ( ((transformComponent.Position.x * viewport.entityScale.x) + (spriteComponent.sourceRectangle.w * viewport.entityScale.x) / 2.0) - (allCameras[viewport.cameraID].frame.getW()) / 2.0);
+    y = ( ((transformComponent.Position.y * viewport.entityScale.y) + (spriteComponent.sourceRectangle.h * viewport.entityScale.y) / 2.0) - (allCameras[viewport.cameraID].frame.getH()) / 2.0);
 
     //x = (transformComponent.Position.x - 10.00);
     //y = (transformComponent.Position.y - 70.00);
